requirements.txt
aiofiles>=24.1.0
aiogram>=3.22.0
aiohappyeyeballs>=2.6.1
aiohttp>=3.12.15
aiosignal>=1.4.0
annotated-types>=0.7.0
anyio>=4.11.0
asyncpg>=0.30.0
attrs>=25.4.0
certifi>=2025.10.5
charset-normalizer>=3.4.4
distro>=1.9.0
frozenlist>=1.8.0
greenlet>=3.2.4
h11>=0.16.0
httpcore>=1.0.9
httpx>=0.28.1
idna>=3.11
jiter>=0.11.1
jsonpatch>=1.33
jsonpointer>=3.0.0
langchain>=0.3.27
langchain-core>=0.3.79
langchain-openai>=0.3.35
langchain-text-splitters>=0.3.11
langgraph>=1.0.1
langgraph-prebuilt>=1.0.1
langgraph-sdk>=0.2.9
langsmith>=0.4.37
magic-filter>=1.0.12
multidict>=6.7.0
openai>=2.6.0
orjson>=3.11.3
ormsgpack>=1.11.0
packaging>=25.0
propcache>=0.4.1
psycopg>=3.2.12
psycopg-pool>=3.2.7
psycopg2>=2.9.11
pydantic>=2.11.10
pydantic-settings>=2.11.0
pydantic_core>=2.33.2
python-dotenv>=1.1.1
PyYAML>=6.0.3
regex>=2025.10.23
requests>=2.32.5
requests-toolbelt>=1.0.0
sniffio>=1.3.1
SQLAlchemy>=2.0.44
tenacity>=9.1.2
tiktoken>=0.12.0
tqdm>=4.67.1
types-requests>=2.32.4.20250913
typing-inspection>=0.4.2
typing_extensions>=4.15.0
urllib3>=2.5.0
xxhash>=3.6.0
yarl>=1.22.0
zstandard>=0.25.0
